+++ src/main/java/org/toilelibre/libe/soundtransform/model/converted/spectrum/SpectrumHelper.java
@@ -2,9 +2,9 @@ package org.toilelibre.libe.soundtransform.model.converted.spectrum;
 
 public interface SpectrumHelper<T> {
 
-    int f0 (Spectrum<T> fs, int i);
+    float f0 (Spectrum<T> fs, int i);
 
-    int freqFromSampleRate (int freq, int sqr2length, int sampleRate);
+    float freqFromSampleRate (float freq, int sqr2length, int sampleRate);
 
     int getMaxIndex (Spectrum<T> fscep, int i, int sampleRate);
