+++ src/test/java/org/toilelibre/libe/soundtransform/TestDetectNotes.java
@@ -22,15 +22,15 @@ public class TestDetectNotes extends SoundTransformTest {
         final int value = 200;
         final int twopercents = (int) (((value * 5.0) / 100) - 1);
         final List<String> messages = new LinkedList<String> ();
-        final int [] t = new int [2000];
+        final float [] t = new float [2000];
         for (int i = 200 ; i < 600 ; i++) {
-            t [i] = (int) ((value + (Math.random () * twopercents)) - (twopercents / 2));
+            t [i] = (float) ((value + (Math.random () * twopercents)) - (twopercents / 2));
         }
         for (int i = 800 ; i < 1000 ; i++) {
-            t [i] = (int) ((value + (Math.random () * twopercents)) - (twopercents / 2));
+            t [i] = (float) ((value + (Math.random () * twopercents)) - (twopercents / 2));
         }
         for (int i = 1100 ; i < 1600 ; i++) {
-            t [i] = (int) ((value + (Math.random () * twopercents)) - (twopercents / 2));
+            t [i] = (float) ((value + (Math.random () * twopercents)) - (twopercents / 2));
         }
         $.create (ImportPackService.class).setObservers (new Slf4jObserver ()).importPack ($.select (Library.class), "default", Thread.currentThread ().getContextClassLoader ().getResourceAsStream ("defaultPack.json"));
         new ShapeSoundTransformation ("default", "simple_piano", t, 200000, 2, 44100).setObservers (new Slf4jObserver (), new Observer () {
